Refactor: move filter part of FindCommandParser and StatsCommandParser to ParserUtil::filterByKeyword
