+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -1,15 +1,18 @@
 package seedu.address.logic.parser;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.logic.parser.CliSyntax.*;
 
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.*;
+import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
+import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.flashcard.Face;
+import seedu.address.model.flashcard.FlashcardContainsKeywordsPredicate;
 import seedu.address.model.tag.Tag;
 
 /**
@@ -85,5 +88,53 @@ public class ParserUtil {
             faceSet.add(parseFace(faceText));
         }
         return faceSet;
+    }
+
+    /**
+     * Parses the given {@code String} of arguments and returns a FlashcardContainsKeywordsPredicate
+     * object.
+     *
+     * @throws ParseException if the user input does not conform the expected format
+     */
+    static FlashcardContainsKeywordsPredicate filterByKeyword(String args) throws ParseException {
+        ArgumentMultimap argMultimap;
+        argMultimap = ArgumentTokenizer.tokenize(args, PREFIX_FRONT_FACE, PREFIX_BACK_FACE, PREFIX_TAG);
+
+        if (!arePrefixesPresent(argMultimap, PREFIX_FRONT_FACE) && !arePrefixesPresent(argMultimap, PREFIX_BACK_FACE)
+                && !arePrefixesPresent(argMultimap, PREFIX_TAG) || !argMultimap.getPreamble().isEmpty()) {
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
+        }
+
+        Set<Face> frontFaceKeywordSet = ParserUtil.parseFaces(argMultimap.getAllValues(PREFIX_FRONT_FACE));
+        Set<Face> backFaceKeywordSet = ParserUtil.parseFaces(argMultimap.getAllValues(PREFIX_BACK_FACE));
+        Set<Tag> tagKeywordSet = ParserUtil.parseTags(argMultimap.getAllValues(PREFIX_TAG));
+
+        ArrayList<String> frontFaceKeywords = new ArrayList<>();
+        ArrayList<String> backFaceKeywords = new ArrayList<>();
+        ArrayList<String> tagKeywords = new ArrayList<>();
+
+        for (Face frontFace : frontFaceKeywordSet) {
+            String[] frontFaceTextSplit = frontFace.text.split("\\s+");
+            frontFaceKeywords.addAll(Arrays.asList(frontFaceTextSplit));
+        }
+
+        for (Face backFace : backFaceKeywordSet) {
+            String[] backFaceTextSplit = backFace.text.split("\\s+");
+            backFaceKeywords.addAll(Arrays.asList(backFaceTextSplit));
+        }
+
+        for (Tag tag : tagKeywordSet) {
+            tagKeywords.add(tag.tagName);
+        }
+
+        return new FlashcardContainsKeywordsPredicate(frontFaceKeywords, backFaceKeywords, tagKeywords);
+    }
+
+    /**
+     * Returns true if none of the prefixes contains empty {@code Optional} values in the given
+     * {@code ArgumentMultimap}.
+     */
+    private static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix... prefixes) {
+        return Stream.of(prefixes).allMatch(prefix -> argumentMultimap.getValue(prefix).isPresent());
     }
 }
