+++ src/main/java/seedu/address/logic/parser/StatsCommandParser.java
@@ -1,19 +1,7 @@
 package seedu.address.logic.parser;
 
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_BACK_FACE;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_FRONT_FACE;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
-
-import java.util.ArrayList;
-import java.util.Set;
-import java.util.stream.Stream;
-
 import seedu.address.logic.commands.StatsCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.flashcard.Face;
-import seedu.address.model.flashcard.FlashcardContainsKeywordsPredicate;
-import seedu.address.model.tag.Tag;
 
 /**
  * Parses input arguments and creates a new StatsCommand object
@@ -30,50 +18,9 @@ public class StatsCommandParser implements Parser<StatsCommand> {
             return new StatsCommand();
         }
 
-        ArgumentMultimap argMultimap =
-                ArgumentTokenizer.tokenize(args, PREFIX_FRONT_FACE, PREFIX_BACK_FACE, PREFIX_TAG);
-
-        if (!arePrefixesPresent(argMultimap, PREFIX_FRONT_FACE) && !arePrefixesPresent(argMultimap, PREFIX_BACK_FACE)
-                && !arePrefixesPresent(argMultimap, PREFIX_TAG) || !argMultimap.getPreamble().isEmpty()) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, StatsCommand.MESSAGE_USAGE));
-        }
-
-        Set<Face> frontFaceKeywordSet = ParserUtil.parseFaces(argMultimap.getAllValues(PREFIX_FRONT_FACE));
-        Set<Face> backFaceKeywordSet = ParserUtil.parseFaces(argMultimap.getAllValues(PREFIX_BACK_FACE));
-        Set<Tag> tagKeywordSet = ParserUtil.parseTags(argMultimap.getAllValues(PREFIX_TAG));
-
-        ArrayList<String> frontFaceKeywords = new ArrayList<>();
-        ArrayList<String> backFaceKeywords = new ArrayList<>();
-        ArrayList<String> tagKeywords = new ArrayList<>();
-
-        for (Face frontFace : frontFaceKeywordSet) {
-            String[] frontFaceTextSplit = frontFace.text.split("\\s+");
-            for (String frontFaceKeyword : frontFaceTextSplit) {
-                frontFaceKeywords.add(frontFaceKeyword);
-            }
-        }
-
-        for (Face backFace : backFaceKeywordSet) {
-            String[] backFaceTextSplit = backFace.text.split("\\s+");
-            for (String backFaceKeyword : backFaceTextSplit) {
-                backFaceKeywords.add(backFaceKeyword);
-            }
-        }
-
-        for (Tag tag : tagKeywordSet) {
-            tagKeywords.add(tag.tagName);
-        }
-
-        return new StatsCommand(
-                new FlashcardContainsKeywordsPredicate(frontFaceKeywords, backFaceKeywords, tagKeywords));
+        return new StatsCommand(ParserUtil.filterByKeyword(args));
     }
 
-    /**
-     * Returns true if none of the prefixes contains empty {@code Optional} values in the given
-     * {@code ArgumentMultimap}.
-     */
-    private static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix... prefixes) {
-        return Stream.of(prefixes).allMatch(prefix -> argumentMultimap.getValue(prefix).isPresent());
-    }
+
 
 }
